Organize sources by language
