+++ .vscode/tasks.json
@@ -4,7 +4,7 @@
     {
       "label": "Run all datasets",
       "type": "shell",
-      "command": "python run_all_datasets.py --verbose",
+      "command": "python Python/run_all_datasets.py --verbose",
       "problemMatcher": "$pytest"
     },
     {
