+++ MATLAB/validate_pipeline_summary.m
