+++ MATLAB/validate_triad_summary.m
