+++ Python/auto_plots.py
