+++ Python/fusion_single.py
