+++ Python/generate_summary.py
