+++ Python/plot_compare_all.py
