+++ Python/run_triad_only.py
