+++ Python/scripts/random_fractal.py
