+++ Python/scripts/validate_filter.py
