+++ Python/validate_with_truth.py
