+++ tests/test_average_rotation.py
@@ -1,7 +1,7 @@
 import os, sys
 sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
 import numpy as np
-from GNSS_IMU_Fusion import average_rotation_matrices
+from Python.GNSS_IMU_Fusion import average_rotation_matrices
 
 def test_average_rotation_identity():
     R1 = np.eye(3)
